moved part SVG components in separate file
